backend/pyproject.toml
[project]
name = "mood-based-music-recommendation-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "spotipy>=2.26.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.41.0",
]
